tools/stack-dashboard/pyproject.toml
[project]
name = "stack-dashboard"
version = "0.1.0"
description = "Reusable Kubernetes stack dashboard with Textual TUI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.89.1",
    "kubernetes>=31.0.0",
    "pyyaml>=6.0.2",
    "pyperclip>=1.9.0",
    "rich>=13.9.4",
]

[project.optional-dependencies]
dev = [
    "textual-dev>=1.7.0",
]

[project.scripts]
stack-dashboard = "stack_dashboard.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stack_dashboard"]

[tool.ruff]
line-length = 100
target-version = "py311"
